requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
motor==2.5.1
pydantic==1.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
twilio==7.10.0
httpx==0.23.0
openpyxl==3.1.2
python-dotenv==1.0.1
